Divides the code into modules
