+++ index.html
@@ -10,7 +10,7 @@
 </head>
 
 <body>
-  <script src="/js/main.js"></script>
+
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
@@ -227,6 +227,9 @@
       </div>
     </section>
   </template>
-
+  <script src="/js/main.js" type="module"></script>
+  <script src="/js/util.js" type="module"></script>
+  <script src="/js/create-comments.js" type="module"></script>
+  <script src="/js/create-photos.js" type="module"></script>
 </body>
 </html>
